Authentication
Created the login and register pages, now the user has to be logged in to create new listings and can only update the listings that are owned by them.

+++ app/Http/Controllers/ListingController.php
@@ -42,6 +42,7 @@ class ListingController extends Controller
         if ($request->hasFile('logo')) {
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
+        $formFields['user_id'] = auth()->id();
 
         Listing::create($formFields);
 
@@ -55,6 +56,11 @@ class ListingController extends Controller
 
     public function update(Request $request, Listing $listing)
     {
+
+        if($listing->user_id != auth()->id()) {
+            abort(403,'Unauthorized Action!');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -76,8 +82,16 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing)
     {
+        if($listing->user_id != auth()->id()) {
+            abort(403,'Unauthorized Action!');
+        }
+
         $listing->delete();
-        
+
         return redirect('/')->with('message', 'Listing deleted successfully!');
+    }
+
+    public function manage(){
+        return view('listings.manage',['listings'=>auth()->user()->listings()->get()]);
     }
 }
